Changes for Lector - add profile page and edit profile page

+++ Services/LectorService.cs
@@ -41,34 +41,38 @@ namespace MMUniGraduation.Services
                 currLector.Bio = input.Bio;
             }
 
-            if (lectorPhoto != null)
+            if (input.Photos != null)
             {
-                //add delete from folder
-                _db.Images.Remove(lectorPhoto);
 
-                await _db.SaveChangesAsync();
-            }
-
-            foreach (var photo in input.Photos)
-            {
-                var extension = Path.GetExtension(photo.FileName).TrimStart('.');
-                var wwwrootPath = _webHost.WebRootPath;
-
-                if (!this.allowedExtensions.Any(x => extension.EndsWith(x)))
+                if (lectorPhoto != null)
                 {
-                    throw new Exception($"Invalid image extension {extension} !");
+                    //add delete from folder
+                    _db.Images.Remove(lectorPhoto);
+
+                    await _db.SaveChangesAsync();
                 }
 
-                var dbImage = new Image
+                foreach (var photo in input.Photos)
                 {
-                    Extension = extension,
-                };
+                    var extension = Path.GetExtension(photo.FileName).TrimStart('.');
+                    var wwwrootPath = _webHost.WebRootPath;
 
-                currLector.Photos.Add(dbImage);
+                    if (!this.allowedExtensions.Any(x => extension.EndsWith(x)))
+                    {
+                        throw new Exception($"Invalid image extension {extension} !");
+                    }
 
-                var physicalPath = $"{wwwrootPath}/img/{dbImage.Id}.{extension}";
-                await using Stream fileStream = new FileStream(physicalPath, FileMode.Create);
-                await photo.CopyToAsync(fileStream);
+                    var dbImage = new Image
+                    {
+                        Extension = extension,
+                    };
+
+                    currLector.Photos.Add(dbImage);
+
+                    var physicalPath = $"{wwwrootPath}/img/{dbImage.Id}.{extension}";
+                    await using Stream fileStream = new FileStream(physicalPath, FileMode.Create);
+                    await photo.CopyToAsync(fileStream);
+                }
             }
 
             await _db.SaveChangesAsync();
